Removes extra test class [skip ci]
